Break into pieces signalling/room.rs [run ci]
